# api/openapi.yaml
openapi: 3.0.3
info:
  title: go-mumble-server REST API
  version: 0.1.0
  description: Management API for the Mumble voice chat server
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from POST /api/v1/auth/login
paths:
...
  /api/v1/servers/{id}/users/{sessionId}/kick:
    post:
      summary: Kick connected user
      operationId: kickUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '204':
          description: Kicked
        '404':
          description: User not found
  /api/v1/servers/{id}/users/{sessionId}/mute:
    post:
      summary: Mute or unmute connected user
      operationId: muteUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [mute]
              properties:
                mute: { type: boolean }
      responses:
        '204':
          description: Updated
        '404':
          description: User not found
  /api/v1/servers/{id}/users/{sessionId}/ban:
    post:
      summary: Ban and kick connected user
      operationId: banUser
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sessionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '204':
          description: Banned and kicked
        '404':
          description: User not found
  /api/v1/servers/{id}/channels:
    get:
      summary: Get channel tree
      operationId: getChannels
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Channel tree
    post:
      summary: Create channel
      operationId: createChannel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                parent_id: { type: integer }
                name: { type: string }
                description: { type: string }
                position: { type: integer }
                is_temporary: { type: boolean }
                max_users: { type: integer }
      responses:
        '201':
          description: Created
  /api/v1/servers/{id}/channels/{channelId}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: channelId
        in: path
        required: true
        schema:
          type: integer
    patch:
      summary: Update channel
      operationId: updateChannel
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                position: { type: integer }
                max_users: { type: integer }
                is_temporary: { type: boolean }
                links: { type: array, items: { type: integer } }
      responses:
        '204':
          description: Updated
    delete:
      summary: Delete channel
      operationId: deleteChannel
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Deleted
...